Extract `Printer` classes
