+++ lib/index.js
@@ -4,7 +4,6 @@ const { preprocess, traverse } = require('@glimmer/syntax');
 const Minimatch = require('minimatch').Minimatch;
 const getConfig = require('./get-config');
 const stripBom = require('strip-bom');
-const chalk = require('chalk');
 let path = require('path');
 
 function buildErrorMessage(moduleId, error) {
@@ -189,100 +188,6 @@ class Linter {
 
     return messages;
   }
-
-  logLintingError(name, moduleName, message) {
-    this._messages.push({
-      rule: name,
-      moduleId: moduleName,
-      message,
-    });
-
-    this.console.log(message);
-  }
-
-  static errorsToMessages(filePath, errors, options) {
-    errors = errors || [];
-    options = options || {
-      verbose: false,
-    };
-
-    if (errors.length === 0) {
-      return '';
-    }
-
-    let errorsMessages = errors.map(error => this._formatError(error, options)).join('\n');
-
-    return `${chalk.underline(filePath)}\n${errorsMessages}\n`;
-  }
-
-  static _formatError(error, options) {
-    let message = '';
-
-    let line = error.line === undefined ? '-' : error.line;
-    let column = error.column === undefined ? '-' : error.column;
-
-    message += chalk.dim(`  ${line}:${column}`);
-
-    if (error.severity === WARNING_SEVERITY) {
-      message += `  ${chalk.yellow('warning')}`;
-    } else {
-      message += `  ${chalk.red('error')}`;
-    }
-
-    message += `  ${error.message}  ${chalk.dim(error.rule)}`;
-
-    if (options.verbose) {
-      message += `\n${error.source}`;
-    }
-
-    return message;
-  }
-
-  reportGitHubActionAnnotations(errors) {
-    if (!process.env.GITHUB_ACTIONS || process.env.DISABLE_GITHUB_ACTIONS_ANNOTATIONS) {
-      return;
-    }
-
-    let files = Object.keys(errors);
-    for (let file of files) {
-      let fileErrors = errors[file];
-      for (let error of fileErrors) {
-        let line = error.line;
-        let col = error.column;
-
-        let annotation = Linter._formatGitHubActionAnnotation(error.message, { file, line, col });
-        this.console.log(annotation);
-      }
-    }
-  }
-
-  static _formatGitHubActionAnnotation(message, options = {}) {
-    message = message || '';
-
-    let output = '::error';
-
-    let outputOptions = Object.keys(options)
-      .map(key => `${key}=${escape(String(options[key]))}`)
-      .join(',');
-
-    if (outputOptions) {
-      output += ` ${outputOptions}`;
-    }
-
-    return `${output}::${escapeData(message)}`;
-  }
-}
-
-function escapeData(s) {
-  return s.replace(/\r/g, '%0D').replace(/\n/g, '%0A');
-}
-
-function escape(s) {
-  return s
-    .replace(/\r/g, '%0D')
-    .replace(/\n/g, '%0A')
-    .replace(/]/g, '%5D')
-    .replace(/;/g, '%3B');
 }
 
 module.exports = Linter;
