+++ test/unit/github-actions-test.js
@@ -1,18 +1,21 @@
 'use strict';
 
-const Linter = require('../../lib/index');
-
-const format = Linter._formatGitHubActionAnnotation;
+const Printer = require('../../lib/printers/github-actions');
 
 describe('GitHub Actions Annotations', function() {
   describe('_formatGitHubActionAnnotation', () => {
+    let printer;
+    beforeEach(() => {
+      printer = new Printer();
+    });
+
     test('formats error annotations', () => {
-      let output = format('foo', { file: 'bar.hbs', line: 4, col: 5 });
+      let output = printer._formatAnnotation('foo', { file: 'bar.hbs', line: 4, col: 5 });
       expect(output).toEqual('::error file=bar.hbs,line=4,col=5::foo');
     });
 
     test('escapes correctly', () => {
-      let output = format('wow!\r\nthis is a multiline\nmessage!!', {
+      let output = printer._formatAnnotation('wow!\r\nthis is a multiline\nmessage!!', {
         file: 'semi;colon.hbs',
         line: 4,
         col: 5,
